+++ CHANGELOG.md
@@ -7,6 +7,7 @@ unreleased
 - `dropinInstance.teardown` will return a promise if no callback is provided
 - `dropin.create` will return a promise if no callback is provided
 - Fix error thrown in console when removing fields with card overrides
+- Fix bug where Drop-in would not finish loading if inside a hidden div
 
 1.3.1
 -----