+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 
 unreleased
 ----------
+- Add `paymentOptionSelected` event
 - `dropinInstance.requestPaymentMethod` will return a promise if no callback is provided
 - `dropinInstance.teardown` will return a promise if no callback is provided
 - `dropin.create` will return a promise if no callback is provided