+++ CHANGELOG.md
@@ -3,9 +3,12 @@ CHANGELOG
 
 unreleased
 ----------
+- Add `paymentOptionSelected` event
 - `dropinInstance.requestPaymentMethod` will return a promise if no callback is provided
 - `dropinInstance.teardown` will return a promise if no callback is provided
 - `dropin.create` will return a promise if no callback is provided
+- Fix error thrown in console when removing fields with card overrides
+- Fix bug where Drop-in would not finish loading if inside a hidden div
 
 1.3.1
 -----